feat: start the core vm implementation
We implement a virtual machine to execute bytecode compiled from JS
source code. This is done for two reasons, first we want a baseline
for our future JIT and second we want to be able to add support for
tracing execution without having to implement a tree walk interpreter
which we might also implement in the future.