feat: add support for function declarations and calls
This is a rather ugly hack, makes me question C++ for this project
either way we fumbled by using JSBasicValue as default for all
return types for expressions.

Refactor will rewrite the expression class to return a shared_ptr
to JSValue, we need it to be shared because it needs to be refcounted.