feat: start refactoring the AST visitor
The goal of the refactor is to add support for the BytecodeCompiler
and the ASTOptimizer. Since both don't return values when visiting
expressions, we refactor the visitor to return void and use a value
stack and expression stack to handle evaluation in the interpreter
and AST node re-write in the optimizing compiler.